Starting to stub out graphics.
First pass, we will just use what SDL2 itself provides us, which
is basically drawing bitmap images.  That's fine.

SDL2_gfx provides stuff for lines, shapes, etc, which it'd be nice
to have someday, but not necessary just yet.  Also that's all in
pure software it seems, so someday we may or may not want to do hardware-
accelerated versions.  (Which is a huge headache.)