Resources system now loads from zipfiles.
It's a bit janky still.  Primarily, just opening a file
from a zipfile is a mutable operation.  Which is irritating.

It works fine really, but it feels like it makes a lot of the
functions that only ever load stuff from files be mutable when
they don't actually mutate anything (except the position of the
cursor in the zip file, which I don't care about.)