Updated filesystem to look in the user directory as well
Which means we can write files now!  Also some general cleanup
and some streamlining of config file loading and saving.

The game's "id", which is used in the user directory path, is now
no longer part of the Conf object, but rather is passed to Game::new().
Because the obvious place to put a config file is... in the user directory.
And if we serialize and write a Conf object, it *has* to be in the user
directory.

So.