Okay it appears to work
But only if you do some nasty hacking to make gfx_window_sdl use the
right version which depends on gfx_core 0.6 which wasn't updated properly
in gfx 0.14, OR you have OpenGL >=4 because a bug in gfx 0.13 will make it
crash if you don't have tesselation support.

So I'm going to leave this branch here, but ignore this until gfx 0.15
comes out, in the hopes that it's less inconsistent.