Fixed strange and terrible opengl driver bugs.
Apparently nvidia drivers magically don't like it when you ask for
the color space sdl asks for by default, which is r3g3b2.  And so
they feel free to do mysterious things like refuse to create a window
the first time you request it, then happily create a window afterwards
even if you request OpenGL 2.5, unless the first window you request is
OpenGL 1 or 2, at which point it works fine but who knows what the heck
it's giving you because it will happily lie.

This happened on Linux, confirmed across multiple people and graphics
cards.  I'd submit a bug to libsdl but I can't be arsed to make a
repro in C.