Changed TTF font rendering a little.
Basically, if V is the amount of the pixel covered by a glyph, as
given by Rusttype, then that pixel's color should be [C,C,C,V].
We were doing it as [V,V,V,V] which resulted in darker boundaries
in the text, so you would get 'fringes' when, say, drawing white text
on a white background.

Resolves issue #85.

Also updated changelog.