Starting to play around with exposing gfx-rs rendering to users.
It's very preliminary just yet but appears to work in the cube.rs
example.  At least, setting the clear color works.  The actual cube
drawing is not working but probably only because of something getting
mixed up in the projection transform, which is always what happens.

The drawing order for mixing ggez and gfx is a little fiddly, since
graphics::present() ALWAYS flushes the ggez Encoder and does the
window swap in one go.  So it will always draw any ggez stuff on
top of whatever else you've drawn.

However, I THINK if we expose the ggez Encoder to the user, nothing's
stopping them from using it with whatever pipeline data and PSO they
want, so they should be able to use it to interleave their drawing
and ggez drawing freely.  :D