Experimental conversion to use nalgebra types.
I have a couple annoyances:

First, there's no methods to convert points to/from arrays, which would
be very useful to have.

Second, EVERYTHING IS DIFFERENT AND IT SUCKS RAAAAAAAH okay not really,
it's fine.

Examples are not converted and there's going to need to be some ergonomic
tests and such, but...