Working on making resizing screen work.
We're getting close...  the projection is correct, I think, and it
resizes the viewport properly, and the coordinates to set_screen_rect()
have the origin in the upper-left.  Need to a) update the graphics_settings
example to show this off better (behavior is different whether the
set_screen_coordinates() is called on resize or not, should make a toggle
to demonstrate), and b) go through everywhere else a Rect is used and
make the coordinates correct there.

Also make sure all the other examples are correct.
Whew!