Moved window title and icon into WindowSetup
Technically they can be set at runtime so WindowMode would be better
but a) we don't expose that functionality yet and nobody's complained
and b) we have no good way of setting *part* of a WindowMode and these
operations are probably rare so forcing you to handle them in every
call to `set_mode()` sounds inconvenient.

Also updated Astroblasto to use a ContextBuilder, which exposed a
couple minor hiccups that have been smoothed over.