A lot of color mongling for SpriteBatch'es.
This would be *slightly* easier without global colors.  Though, admittedly, still not trivial.

Now, if you have a global color set with `graphics::set_color()`, the SpriteBatch will use that.  If you have a color specified
in `DrawParam` when you draw the sprite batch, it will use that instead.  If you have a color specified in the `DrawParam` when
you insert/set the sprite, then the sprite batch will use *that* instead.

Whew.