Replaced all gfx-rs types that depend on Srgba8 with untyped versions
Starts to address #299.

gfx-rs does its best to make `RenderTargetView`, `Texture` and other
such things strongly-typed at compile time with generics.  The
inconvenience here is that the user then can't have any option to set
the color format for these things at runtime.

So instead we rip out all the type-safe types and replace them with
the `Raw` equivalents, which take type info at runtime.  The API is
a little bit different for these types though, and there are some methods
like `Encoder::clear()` that don't have direct equivalents (until I make
a PR to add them).  In general it's giving a slightly less-used portion
of the gfx-rs API a good workout.

This is also just an initial proof-of-concept, so the Srgba8 color format
and some related types are still hard-wired in a few places, marked with
BUGGO comments.  If this functions at all then we can start tackling those
one by one.

It would be way easier to just make the `GraphicsContext` and associated
types generic on the color format, but that still results in it being
fixed at compile-time unless we dive into outright trait objects.  I'm
reluctant to do that just for this 'cause ggez isn't built that way,
but given ggez's assumption that all color formats everywhere are always
the same, then it might not be a bad idea.