Turns out making DrawParam be *only* a matrix is inconvenient.
Some of the matrix components (scale, shear and rotation) overlap,
and extracting them is non-trivial.  Composing them means the order in
which you do things matters.  So we have to keep track of things
anyway.  SO, while we can turn a DrawParam into a Matrix4, turning
a Matrix4 into a DrawParam is non-trivial, so we now have a separate
type that does this.  :|