Let's try the "generics EVERYWHERE" approach.
This will make Image, Shader, Context etc. all
generic on the graphics color format.

I don't like it but it seems simple and achievable
for now, and might let us ponder the problem from
a new point of view.

The alternative seems to be making EVERYTHING a trait
object to hide all the format stuff, which I don't
like, soooo....