Separated tests into headless and non-headless.
Only tests with "headless" in the name will be run by CI suites,
which is a somewhat crude but also very simple way of separating
the two.  Firmly in "why didn't we think of this before?"
territory.

Dunno how to make Travis give a custom command line to cargo yet
but I'm on a plane so can't check.