Add docs to indicate that primitive draws are inefficient
The snake example is referenced from a youtube video and a blog post, so
I added documentation there because that's likely one of the first
places a new developer will interact with ggez and base projects on.

I noted in the graphics package that the draws are different from the
love project that it's based on since primitive draw calls seem to be
batched, whereas ggez does *not* do automatic batching. There may be
more places in the source where such notes are useful.

closes #440