s/fn get_/fn / in graphics package
I left get_or_insert because that seemed to fit with how the standard
library did things, namely Option.get_or_insert.

I also left pub fns because there's a get_transform and transform and I
couldn't decide on a better naming scheme.

This modifies the Drawable and ShaderHandle traits, so I added a note in
the changelog of the breaking change.