Rename ALL getters from `get_foo()` to `foo()`.
See issue #424 for discussion.

So far I rather like this change but it's going to live in its own
branch for a little because it's VERY far-reaching.

I tried having builder structs use `with_foo()` instead of `foo()`
but I find I like that rather less, even if it conflates the
meanings a little.