Update README.md
Added a suggestion to read the `hello ggez` guide for new users. It's better to let the to-be users immediately know that there exists a 'getting started' tutorial such as this. It may be hard to find otherwise, as it requires manually looking at the `docs/` folder to find it.