Messing around with sound API stuff.
It's far from perfect (doesn't even build)
but I'm just trying to flesh out the thoughts
in my head.  Contrast with #591.

To make this work we'd probably have to pull the
`Sink` out of the `Source` and `SpatialSource` types,
make everything else in `SourceImpl` be a new struct
shared between `Source` and `SpatialSource`, and
refactor lots of simple methods.  Not the best but
not the worst maybe?

This is another one of those cases where the lines you
have to cut along to break things up nicely are not where
you think they are, it seems.