Resolve #587.
I'm sick of this and want to just move on with life.  Any other
issue with hidpi anything is going to be closed as "dependency bug"
until winit eventually switches to physical coordinates by default,
which admittedly they're probably going to.

As a consequence, Apple platforms are no longer supported.
See https://drewdevault.com/2017/10/26/Fuck-you-nvidia.html but
search-and-replace NVidia with Apple.  It's the exact same sort
of problem and I'm done with it.

I'm sorry.