Make uv consistent in vertex builder
Using the position as uv for the FillVertex builder is pretty arbitrary and not consistent with the StrokeVertex builder. 
I'm trying to mix basic shapes and textures shapes in one big mesh (with one image acting as an atlas) and having a consistent uv for basic shapes (not made with MeshBuilder::raw) is important.