Cleanup remains of matrix stack.
One is tempted to allow users to access
GraphicsContext::set_global_mvp() somehow, but
really it's best if we don't because then we have
to juggle their state with our state in spritebatch
and meshbatch.  So really, it would just become
a one-item stack anyway.  Best if we just let
them build everything themselves and do the
math on `DrawParam`'s without any more
sneaky global state.