Remove `TimeUpdateError` :-(
Result is the wrong tool for the job, because it can trivially
be returned with `?` instead of being handled.  If you do that it
will get passed up to the event loop, which will log it (silently
by default) and then terminate.  This leads to Mysterious Things like
your game starting, lagging a moment on the first frame as it tries
to load everything or create the graphics context, and then silently
quitting.

Still need to have some kind of functionality like this, but I guess
we should do it differently.