Remove backticks from GenerativeArt.md section
Backticks were being used in a section header to format `ggez::graphics`
as code. While this looked good, the skeptic automated markdown tests
do not parse backticked strings as names, so the section header was
parsed as an empty string. This resulted in test names like this one:
```
```
The `__` was flagged by `cargo check` as poorly typed snake_case.

The solution is to simply remove the backticks.