Make most of the `filesystem` functions take `&Context`.
Most of these functions do not need to take `&mut Context` and can simply
take `&Context` instead. This is helpful since there are some scenarios
where you want `&Context`s, such as when passing the same context to a
bunch of closures that all want to read a file and don't need mutability.