Add logical_size to WindowMode and use it when set to add DPI aware window sizing
- Fixes https://github.com/ggez/ggez/issues/1091

The default width/height ggez uses is converted into a winit's PhysicalSize and used in two places:
  - The WindowBuilder for creating the window initially
  - set_window_mode which is called shortly after the window is first shown (1 frame from what I can tell)

This works great, except on high DPI machines, where 800x600 is exceptionally small.

winit has a LogicalSize which is correctly scaled by the current screen's DPI. However, as we do not want to break the common/easy case of width/height, it is being added as a optional argument on WindowMode (width/height is ignored when set to Some).

To add this to WindowMode, winit must include the serde feature as WindowMode is marked Serialize.

Usages of `dpi::PhysicalSize` that access conf.width/conf.height are now incorrect, so I added a crate public to create actual_size on WindowMode to prevent duplication.

One error case in resizing has been converted into an assert for simplicity, it only appears to be possible if the Window is set to a larger than 1x1 size and then resized to 0x0.